+++ src/video/yuv2rgb/yuv_rgb_common.h
@@ -1,0 +1,13 @@
+#ifndef YUV_RGB_COMMON_H_
+#define YUV_RGB_COMMON_H_
+// Copyright 2016 Adrien Descamps
+// Distributed under BSD 3-Clause License
+
+typedef enum
+{
+    YCBCR_JPEG,
+    YCBCR_601,
+    YCBCR_709
+} YCbCrType;
+
+#endif /* YUV_RGB_COMMON_H_ */
